feat: abstract common trait for FrameAllocator
