+++ src/common/mod.rs
@@ -1,3 +1,4 @@
 pub mod addr;
+pub mod frame_allocator;
 pub mod pagetable;
 pub mod tlb;
